week3_labs/userlogin/requirements.txt
flet>=0.7.0
mysql-connector-python>=8.0.0
